move project files listener to device session
