+++ packages/vscode-extension/src/project/project.ts
@@ -21,7 +21,7 @@ import {
   ZoomLevelType,
 } from "../common/Project";
 import { Logger } from "../Logger";
-import { DeviceInfo, DevicePlatform } from "../common/DeviceManager";
+import { DeviceInfo } from "../common/DeviceManager";
 import { DeviceManager } from "../devices/DeviceManager";
 import { extensionContext } from "../utilities/extensionContext";
 import {
@@ -39,8 +39,6 @@ import { findAndSetupNewAppRootFolder } from "../utilities/findAndSetupNewAppRoo
 import { getLaunchConfiguration } from "../utilities/launchConfiguration";
 import { DeviceSessionsManager, DeviceSessionsManagerDelegate } from "./DeviceSessionsManager";
 import { DEVICE_SETTINGS_DEFAULT, DEVICE_SETTINGS_KEY } from "../devices/DeviceBase";
-import { watchProjectFiles } from "../utilities/watchProjectFiles";
-import { throttleAsync } from "../utilities/throttle";
 
 const PREVIEW_ZOOM_KEY = "preview_zoom";
 const DEEP_LINKS_HISTORY_KEY = "deep_links_history";
@@ -48,7 +46,6 @@ const DEEP_LINKS_HISTORY_KEY = "deep_links_history";
 const DEEP_LINKS_HISTORY_LIMIT = 50;
 
 const MAX_RECORDING_TIME_SEC = 10 * 60; // 10 minutes
-const CACHE_STALE_THROTTLE_MS = 10 * 1000; // 10 seconds
 
 export class Project implements Disposable, ProjectInterface, DeviceSessionsManagerDelegate {
   private applicationContext: ApplicationContext;
@@ -112,42 +109,7 @@ export class Project implements Disposable, ProjectInterface, DeviceSessionsMana
         }
       })
     );
-
-    this.disposables.push(watchProjectFiles(this.onProjectFilesChanged));
   }
-
-  private onProjectFilesChanged = throttleAsync(async () => {
-    const sessions = this.deviceSessionsManager.deviceSessions.values().toArray();
-    const platforms = _.uniq(sessions.map((session) => session.platform));
-    const launchConfig = getLaunchConfiguration();
-    for (const platform of platforms) {
-      const hasCachedBuild = this.applicationContext.buildCache.hasCachedBuild(
-        platform,
-        this.appRootFolder
-      );
-      const platformKey: "ios" | "android" = platform === DevicePlatform.IOS ? "ios" : "android";
-      if (hasCachedBuild) {
-        const fingerprint = await this.applicationContext.buildCache.calculateFingerprint({
-          appRoot: this.appRootFolder,
-          env: launchConfig.env,
-          fingerprintCommand: launchConfig.customBuild?.[platformKey]?.fingerprintCommand,
-        });
-        const isCacheStale = await this.applicationContext.buildCache.isCacheStale(
-          fingerprint,
-          platform,
-          this.appRootFolder
-        );
-
-        if (isCacheStale) {
-          sessions
-            .filter((session) => session.platform === platform)
-            .forEach((session) => {
-              session.onCacheStale();
-            });
-        }
-      }
-    }
-  }, CACHE_STALE_THROTTLE_MS);
 
   onDeviceSessionsManagerStateChange(state: DeviceSessionsManagerState): void {
     this.updateProjectState(state);
